refactor: cliente externo
