+++ src/main/java/nttdata/bank/feign/clients/ContaExternaClient.java
@@ -1,13 +1,15 @@
 package nttdata.bank.feign.clients;
 
 
-import nttdata.bank.feign.response.ClienteExterno;
+import nttdata.bank.feign.config.FeignInterceptorConfig;
+import nttdata.bank.feign.response.mockapi.ClienteExterno;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 
 import java.util.Optional;
 
-@FeignClient(name = "cliente-externo-api", url = " https://api.mockapi.com/api/v1")
+@FeignClient(name = "cliente-externo-api", url = "https://api.mockapi.com/api/v1",
+configuration = FeignInterceptorConfig.class)//TRANSFERIR PARA O APPLICATION.PROPERTIES
 public interface ContaExternaClient {
 
     @GetMapping("/cliente-externo")
