+++ src/main/java/nttdata/bank/feign/impl/ClienteExternoServiceImpl.java
@@ -1,7 +1,7 @@
 package nttdata.bank.feign.impl;
 
 import nttdata.bank.feign.clients.ContaExternaClient;
-import nttdata.bank.feign.response.ClienteExterno;
+import nttdata.bank.feign.response.mockapi.ClienteExterno;
 import nttdata.bank.service.adapters.ClienteExternoService;
 import org.springframework.stereotype.Service;
