+++ src/main/java/nttdata/bank/service/adapters/ClienteExternoService.java
@@ -1,6 +1,5 @@
 package nttdata.bank.service.adapters;
 
-import nttdata.bank.feign.response.ClienteExterno;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
@@ -8,5 +7,5 @@ import java.util.Optional;
 @Service
 public interface ClienteExternoService {
 
-    Optional<ClienteExterno> buscarClienteExterno();
+    Optional<?> buscarClienteExterno();
 }
